requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
python-multipart==0.0.20
tensorflow==2.15.0
pillow==10.2.0
numpy==1.26.0 